Restructures premove survey e2e tests
