+++ cypress/integration/office/officeUserShipment.js
@@ -3,9 +3,6 @@ describe('office user finds the move', function() {
   beforeEach(() => {
     cy.signIntoOffice();
   });
-  it('office user enters premove survey', function() {
-    officeUserEntersPreMoveSurvey();
-  });
   it('office user views hhg moves in queue new moves', function() {
     officeUserViewsMoves();
     cy.resetDb();
@@ -36,36 +33,4 @@ function officeUserViewsMoves() {
   cy.location().should(loc => {
     expect(loc.pathname).to.match(/^\/queues\/new\/moves\/[^/]+\/hhg/);
   });
-}
-
-function officeUserEntersPreMoveSurvey() {
-  // Open new moves queue
-  cy.visit('/queues/all');
-  cy.location().should(loc => {
-    expect(loc.pathname).to.match(/^\/queues\/all/);
-  });
-
-  // Find move and open it
-  cy
-    .get('div')
-    .contains('RLKBEM')
-    .dblclick();
-
-  cy.location().should(loc => {
-    expect(loc.pathname).to.match(/^\/queues\/new\/moves\/[^/]+\/basics/);
-  });
-
-  // Click on HHG tab
-  cy
-    .get('span')
-    .contains('HHG')
-    .click();
-  cy.location().should(loc => {
-    expect(loc.pathname).to.match(/^\/queues\/new\/moves\/[^/]+\/hhg/);
-  });
-
-  // Verify that the Estimates section contains expected data
-  cy.get('span').contains('2,000');
-
-  cy.testPremoveSurvey();
 }
